[1015,1020]: Change position of the tabs on my properties page and change styles of the scrollbar (#1022)
* fix: align selected item in select component

* fix(my-properties-page): change position of the tabs

* fix: scrollbar styles are not matching with design
