+++ src/styles/scrollbar.scss
@@ -1,6 +1,7 @@
 ::-webkit-scrollbar {
-  width: 3px;
-  background-color: black;
+  width: 4px;
+  height: 4px;
+  background-color: transparent;
 }
 
 ::-webkit-scrollbar-button {
@@ -8,10 +9,10 @@
 }
 
 ::-webkit-scrollbar-thumb {
-  border-right: 3px solid var(--theme-modal-color);
-  width: 3px;
+  border-radius: 2px;
+  background-color: var(--theme-grey400-color);
 }
 
 ::-webkit-scrollbar-corner {
-  background: rgba(0, 0, 0, 0);
+  background: transparent;
 }
